Code review suggestions
Using the PR's diff and our own patch, compute suggestion comments and
leave them on the PR.

These comments will appear in line, in context, and can be committed
individually or in batch. This can happen on same-origin and fork PRs,
and obviates the need for sibling PRs or copy/pasting `git-am` patches.

Deferred features:

1. What to do if suggestions can't be commented (because the diff is
   disorganized, not on added lines, etc)
2. Listen when users resolve a given suggestions and not re-make it each
   time