Respect resolved comments
Don't emit a suggestion if there's already a resolved one there. This means users
can use that to indicate they don't want to apply the Restyled suggestion, and
we won't continue to make it on every push.

This required switching to GraphQL because the resolved and minimized
attributes are not present in the REST API endpoints.