Avoid fetching completely PR diff as patch
This fails for large PRs, and the `/files` response we already have
contains enough of the diff details to do what we need, which is confirm
that our restyled are contained within added lines of the original PR.

To do this, we introduce a `ChangeFile` type that we can build from the
`/files` response. It can hold hunks of added lines, so we can do our
`.contains` check with it.

It even reads better this way,

```js
changedFile.additions.contains(_)
```

:ok_hand: