feat(setup): use gh
Using the main `install` script is actually not very robust when used
with great frequency. It often fails on 403s due to, I assume, rate
limiting, since there's no auth involved.

Installing locally is a bit of an edge-case anyway, and it's so far been
mostly used on CI, where we have `gh` available to simplify interacting
with releases.