fix(setup-demo): remove ignores in test restyled.yaml
If we're using this action as a form of CI, we want to run. Skipping
renovate, for example, is not desired and in fact harmful.

Using `""` values is a bit of a wart in Restyler config parsing. These
will become singleton lists of an empty glob, due to a quirk of the
`opt-env-conf` `commaSeparatedList` parser; but this value matches
nothing (at least in `ghci` testing), so the behavior will be as
desired.